requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
